requirements.txt
streamlit
python-dotenv
openai
zhipuai
langchain
langchain-text-splitters
pymupdf
rank-bm25
numpy
